move generated deps out of pkg
