+++ pkg/middleware/auth/auth.go
@@ -29,6 +29,7 @@ import (
 	"github.com/theopenlane/core/internal/ent/generated/orgmembership"
 	"github.com/theopenlane/core/internal/ent/generated/personalaccesstoken"
 	"github.com/theopenlane/core/internal/ent/generated/privacy"
+	"github.com/theopenlane/core/internal/ssoenforcement"
 	"github.com/theopenlane/core/pkg/logx"
 	"github.com/theopenlane/core/pkg/metrics"
 	"github.com/theopenlane/core/pkg/permissioncache"
@@ -617,7 +618,7 @@ func isSSOEnforced(ctx context.Context, db *ent.Client, orgID string) (bool, err
 // flow, taking owner, per-user, and per-domain exemptions into account. TFA enforcement is
 // evaluated separately and is not affected by SSO exemption
 func userMustSSO(ctx context.Context, db *ent.Client, orgID, userID string) (bool, error) {
-	in, _, err := sso.LoadEnforcement(ctx, db, orgID, userID, "")
+	in, _, err := ssoenforcement.LoadEnforcement(ctx, db, orgID, userID, "")
 	if err != nil {
 		if ent.IsNotFound(err) {
 			return false, nil
